slight component type refactor
