+++ src/ColorPicker.tsx
@@ -1,7 +1,7 @@
 import { h, Component } from 'preact';
 import { IroColor, IroColorValue, IroColorPickerOptions, iroColorPickerOptionDefaults } from '@irojs/iro-core';
 
-import { IroInputType } from './ComponentBase';
+import { IroInputType } from './ComponentTypes';
 import { IroWheel } from './Wheel';
 import { IroSlider } from './Slider';
 import { createWidget } from './createWidget';
